Rename crud.js to task-service.js
