+++ src/lib/Reusable/FlexibleDayTask.svelte
@@ -46,7 +46,7 @@
   // Assumes `task` is hydrated
   import { grabOffset, activeDragItem, openTaskPopup } from '/src/store'
   import Checkbox from './Checkbox.svelte'
-  import { updateTaskNode } from '/src/db/crud.js'
+  import { updateTaskNode } from '/src/db/task-service.js'
 
   export let task = null
   export let pixelsPerHour = null
