+++ src/lib/Reusable/PhotoTaskElement.svelte
@@ -99,7 +99,7 @@
  import { grabOffset, activeDragItem, openTaskPopup } from '/src/store'
  import { lazyCallable } from '/src/utils/svelteActions.js'
  import { pixelsPerHour } from '/src/lib/Calendar/store.js'
- import { updateTaskNode } from '/src/db/crud.js'
+ import { updateTaskNode } from '/src/db/task-service.js'
 
  export let task = null
