+++ src/lib/Reusable/TaskElement.svelte
@@ -113,7 +113,7 @@
   import { grabOffset, activeDragItem, openTaskPopup } from '/src/store'
   import Checkbox from './Checkbox.svelte'
   import { pixelsPerHour, treesByID } from '/src/lib/Calendar/store.js'
-  import { updateTaskNode } from '/src/db/crud.js'
+  import { updateTaskNode } from '/src/db/task-service.js'
 
   export let task = null
   export let hasCheckbox = false
